Fixed every warning that was possible to fix.

+++ src/model/gameboard.cpp
@@ -10,7 +10,7 @@
 #include "../frost_util.h"
 
 GameBoard::GameBoard(const unsigned int& newWidth, const unsigned int& newHeight, 
-const uint8_t& newTileTypes) : matchesNeeded(0) {
+const unsigned int& newTileTypes) : matchesNeeded(0) {
 
     boardDimensions.width = FrostUtil::ClampInts(newWidth, GameBoardConstants::MIN_WIDTH, GameBoardConstants::MAX_WIDTH);
     boardDimensions.height = FrostUtil::ClampInts(newHeight, GameBoardConstants::MIN_HEIGHT, GameBoardConstants::MAX_HEIGHT);
@@ -37,7 +37,7 @@ bool GameBoard::tryNewGame() {
     }
     
     // Get the number of tiles this board has
-    uint8_t numTiles = static_cast<uint8_t>((boardDimensions.width * boardDimensions.height));
+    unsigned int numTiles = boardDimensions.width * boardDimensions.height;
     matchesNeeded = numTiles / 2;
 
     for(unsigned int i = 0; i < numTiles; ++i) {
@@ -50,14 +50,14 @@ bool GameBoard::tryNewGame() {
         tiles[(numTiles-1) / 2].flags = TileFlags::MATCHED;
     }
 
-    uint8_t firstUnfilledIndex = 0;
-    uint8_t matchesMade = 0;
+    unsigned int firstUnfilledIndex = 0;
+    unsigned int matchesMade = 0;
 
     srand(static_cast<unsigned int>(time(NULL)));
 
     for(unsigned int i = 0 ;; ++i) {
 
-        const uint8_t TILE_ID = (rand() % numTileTypes) + 2;
+        const unsigned int TILE_ID = (rand() % numTileTypes) + 2;
 
         // Seek the first tile that does not have data set
         if(tiles[firstUnfilledIndex].ID) {
@@ -75,7 +75,7 @@ bool GameBoard::tryNewGame() {
 
         // Select Random Tile for matching tile
 
-        uint8_t matchIndex = (firstUnfilledIndex) + (rand() % (numTiles - firstUnfilledIndex)) - 1;
+        unsigned int matchIndex = (firstUnfilledIndex) + (rand() % (numTiles - firstUnfilledIndex)) - 1;
 
         // Now search for the first empty tile, which may even be the one we just
         // generated.
@@ -131,9 +131,9 @@ bool GameBoard::tryNewGame() {
    
     wchar_t outStr[5] = {0};
 
-    uint8_t k = 0;
+    unsigned int k = 0;
 
-    for(uint8_t i = 0; i < numTiles; ++i) {
+    for(unsigned int i = 0; i < numTiles; ++i) {
 
         wsprintf(outStr, L"%02d \0", tiles[i].ID);
